Reworked the thresholding.
When a parent Logger has a loggable appender, calling the approrpriate
log method for any of its children should result in logging to that
appender.  Conversely, when a parent Logger has a *higher* threshold,
calling a log method for its child should also result in logging.

For example, if the root appender is set to ERROR and a child is set to
INFO, calling child.info should result in the root's appenders being
logged to.